# backend/src/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Portfolio Damien Riandiere API",
    "description": "Development server"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "default": false
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ApiSuccess": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "required": true
          },
          "email": {
            "type": "string",
            "required": true,
            "unique": true
          },
          "password": {
            "type": "string",
            "required": true
          },
          "admin": {
            "type": "boolean",
            "required": true
          }
        }
      },
      "UserToken": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "tokens": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string",
                "required": true,
                "unique": true
              },
              "refreshToken": {
                "type": "string",
                "required": true,
                "unique": true
              },
              "expiresIn": {
                "type": "number",
                "required": true
              },
              "refreshExpiresIn": {
                "type": "number",
                "required": true
              }
            }
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "security": [],
        "tags": [
          "auth"
        ],
        "summary": "refresh",
        "operationId": "refreshToken",
        "description": "Refresh a token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccess"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserToken"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}